django/requirements.txt
python-dotenv
djangorestframework 
mysqlclient
django-cors-headers
django-filter
djangorestframework-simplejwt
django-imagekit
